+++ evalml/pipelines/components/__init__.py
@@ -22,10 +22,14 @@ from .transformers import (
     OneHotEncoder,
     RFClassifierSelectFromModel,
     RFRegressorSelectFromModel,
+    PerColumnImputer,
     SimpleImputer,
     StandardScaler,
     FeatureSelector,
     CategoricalEncoder,
+    DropColumns,
+    DropNullColumns,
+    DateTimeFeaturization
     )
 
 from .utils import all_components, handle_component