+++ evalml/pipelines/components/transformers/imputers/__init__.py
@@ -1,3 +1,4 @@
 # flake8:noqa
+from .per_column_imputer import PerColumnImputer
 from .simple_imputer import SimpleImputer