+++ docs/source/changelog.rst
@@ -5,10 +5,17 @@ Changelog
 
 **Future Releases**
     * Enhancements
+        * Added multiclass support for ROC curve graphing :pr:`832`
+        * Added preprocessing component to drop features whose percentage of NaN values exceeds a specified threshold :pr:`834`
+        * Added data check to check for problematic target labels :pr:`814`
+        * Added PerColumnImputer that allows imputation strategies per column :pr:`824`
+        * Added transformer to drop specific columns :pr:`827`
+        * Added preprocessing component to handle DateTime columns featurization :pr:`838`
     * Fixes
     * Changes
     * Documentation Changes
     * Testing Changes
+      * Update the changelog check job to expect the new branching pattern for the deps update bot :pr:`836`
 
 
 **v0.10.0 May 29, 2020**